[feat][#33]mainpage생성 후 carousel component넣기: 일단 라이브러리로 적용한 carousel component적용 Hooks로 직접 만들었던 carousel은 좀 더 작업해보고 다듬어지면 대체할 계획
