+++ client/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { BrowserRouter as Router, Switch, Route} from 'react-router-dom'
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import { client } from './ApolloClient'
 import { ApolloProvider } from 'react-apollo'
 import { LoginPage } from '@/pages/LoginPage'
@@ -7,6 +7,7 @@ import '@/app.scss'
 import '@/styles/reset.scss'
 import '@/styles/base.scss'
 import '@/styles/fonts.scss'
+import { MainPage } from './pages/MainPage'
 
 function App() {
   return (
@@ -16,6 +17,11 @@ function App() {
           <Switch>
             <Route path="/login">
               <LoginPage />
+            </Route>
+          </Switch>
+          <Switch>
+            <Route path="/main">
+              <MainPage />
             </Route>
           </Switch>
         </div>
