+++ client/src/components/Carousel/practice.scss
@@ -1,0 +1,43 @@
+@import 'src/styles/package';
+
+.carousel {
+  position: relative;
+  max-width: 450px;
+  @include flex-center();
+  flex-direction: row;
+  overflow: hidden;
+  touch-action: auto;
+
+  .cur-number {
+    position: absolute;
+    bottom: 1em;
+    right: 1em;
+    color: black;
+    background-color: lightgray;
+    padding: 0.3em;
+    border-radius: 0.5em;
+  }
+
+  .left-arrow-btn {
+    position: absolute;
+    left: 1em;
+    color: white;
+  }
+
+  .images {
+    display: flex;
+    flex-direction: row;
+    transform: translate(-450px);
+
+    img {
+      width: 100%;
+      height: 10rem;
+    }
+  }
+
+  .right-arrow-btn {
+    position: absolute;
+    right: 1em;
+    color: white;
+  }
+}
