--- client/src/components/Carousel/slickStyle.scss
@@ -1,21 +1,0 @@
-@import 'src/styles/package';
-
-.slick-carousel {
-  max-width: 450px;
-
-  .slick-slider {
-    .slick-prev {
-    }
-    .slick-list {
-    }
-    .slick-next {
-    }
-    .slick-dots {
-      bottom: 0px;
-    }
-  }
-  .slick-prev:before,
-  .slick-next:before {
-    display: none;
-  }
-}
