+++ client/src/components/Carousel/style.scss
@@ -1,43 +1,28 @@
 @import 'src/styles/package';
 
-.carousel {
-  position: relative;
+.slick-carousel {
   max-width: 450px;
-  @include flex-center;
-  flex-direction: row;
-  overflow: hidden;
-  touch-action: auto;
+  position: relative;
 
-  .cur-number {
-    position: absolute;
-    bottom: 1em;
-    right: 1em;
-    color: black;
-    background-color: lightgray;
-    padding: 0.3em;
-    border-radius: 0.5em;
-  }
-
-  .left-arrow-btn {
-    position: absolute;
-    left: 1em;
-    color: white;
-  }
-
-  .images {
-    display: flex;
-    flex-direction: row;
-    transform: translate(-450px);
-
-    img {
-      width: 100%;
-      height: 10rem;
+  .slick-slider {
+    .slick-prev {
+    }
+    .slick-list {
+    }
+    .slick-next {
+    }
+    .slick-dots {
+      bottom: 0px;
     }
   }
+  .slick-prev:before,
+  .slick-next:before {
+    display: none;
+  }
 
-  .right-arrow-btn {
+  .current-slide-number {
     position: absolute;
+    top: 1em;
     right: 1em;
-    color: white;
   }
 }
