+++ client/src/pages/LoginPage.tsx
@@ -1,15 +1,10 @@
 import React from 'react'
 import { Login } from '@/components/Login'
-import { Carousel } from '@/components/Carousel'
-
-import { SlickCarousel } from '@/components/Carousel/slick.tsx'
 
 export const LoginPage = () => {
   return (
     <div>
-      <Carousel />
       <Login />
-      <SlickCarousel />
     </div>
   )
 }
