+++ client/src/pages/MainPage/gql.ts
@@ -1,0 +1,1 @@
+import gql from 'graphql-tag'
