+++ client/src/pages/MainPage/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { SlickCarousel } from '@/components/Carousel/index.tsx'
+
+export const MainPage: React.FC = () => {
+  return (
+    <div id="main-page">
+      <SlickCarousel />
+    </div>
+  )
+}
