+++ client/src/pages/MainPage/style.scss
@@ -1,0 +1,5 @@
+@import 'src/styles/package';
+
+#main-page {
+  width: 100%;
+}
